+++ src/renderer/src/components/ui/Skeleton/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Skeleton exports
+ */
+
+export { Skeleton } from './Skeleton'
+export type { SkeletonProps } from './Skeleton'